#1233 refactored the file/block managers to reduce duplicated code
