+++ SoftLayer/fixtures/SoftLayer_Location_Datacenter.py
@@ -1,0 +1,12 @@
+getDatacenters = [
+    {
+        "id": 1441195,
+        "longName": "Dallas 10",
+        "name": "dal10"
+    },
+    {
+        "id": 449494,
+        "longName": "Dallas 9",
+        "name": "dal09"
+    }
+]
