+++ tests/managers/block_tests.py
@@ -8,7 +8,11 @@
 import copy
 import SoftLayer
 from SoftLayer import exceptions
-from SoftLayer import fixtures
+from SoftLayer.fixtures import SoftLayer_Account
+from SoftLayer.fixtures import SoftLayer_Product_Package
+from SoftLayer.fixtures import SoftLayer_Product_Order
+from SoftLayer.fixtures import SoftLayer_Network_Storage
+from SoftLayer.fixtures import SoftLayer_Network_Storage_Allowed_Host
 from SoftLayer import testing
 
 
@@ -83,7 +87,7 @@ class BlockTests(testing.TestCase):
     def test_get_block_volume_details(self):
         result = self.block.get_block_volume_details(100)
 
-        self.assertEqual(fixtures.SoftLayer_Network_Storage.getObject, result)
+        self.assertEqual(SoftLayer_Network_Storage.getObject, result)
 
         expected_mask = 'id,' \
                         'username,' \
@@ -119,7 +123,7 @@ class BlockTests(testing.TestCase):
     def test_list_block_volumes(self):
         result = self.block.list_block_volumes()
 
-        self.assertEqual(fixtures.SoftLayer_Account.getIscsiNetworkStorage,
+        self.assertEqual(SoftLayer_Account.getIscsiNetworkStorage,
                          result)
 
         expected_filter = {
@@ -157,7 +161,7 @@ class BlockTests(testing.TestCase):
                                                storage_type="Endurance",
                                                username="username")
 
-        self.assertEqual(fixtures.SoftLayer_Account.getIscsiNetworkStorage,
+        self.assertEqual(SoftLayer_Account.getIscsiNetworkStorage,
                          result)
 
         expected_filter = {
@@ -197,7 +201,7 @@ class BlockTests(testing.TestCase):
     def test_get_block_volume_access_list(self):
         result = self.block.get_block_volume_access_list(100)
 
-        self.assertEqual(fixtures.SoftLayer_Network_Storage.getObject, result)
+        self.assertEqual(SoftLayer_Network_Storage.getObject, result)
 
         self.assert_called_with(
             'SoftLayer_Network_Storage',
@@ -207,7 +211,7 @@ class BlockTests(testing.TestCase):
     def test_get_block_volume_snapshot_list(self):
         result = self.block.get_block_volume_snapshot_list(100)
 
-        self.assertEqual(fixtures.SoftLayer_Network_Storage.getSnapshots,
+        self.assertEqual(SoftLayer_Network_Storage.getSnapshots,
                          result)
 
         self.assert_called_with(
@@ -218,7 +222,7 @@ class BlockTests(testing.TestCase):
     def test_delete_snapshot(self):
         result = self.block.delete_snapshot(100)
 
-        self.assertEqual(fixtures.SoftLayer_Network_Storage.deleteObject,
+        self.assertEqual(SoftLayer_Network_Storage.deleteObject,
                          result)
 
         self.assert_called_with(
@@ -331,7 +335,7 @@ class BlockTests(testing.TestCase):
         result = self.block.failover_to_replicant(1234, 5678)
 
         self.assertEqual(
-            fixtures.SoftLayer_Network_Storage.failoverToReplicant, result)
+            SoftLayer_Network_Storage.failoverToReplicant, result)
         self.assert_called_with(
             'SoftLayer_Network_Storage',
             'failoverToReplicant',
@@ -343,7 +347,7 @@ class BlockTests(testing.TestCase):
         result = self.block.failback_from_replicant(1234)
 
         self.assertEqual(
-            fixtures.SoftLayer_Network_Storage.failbackFromReplicant, result)
+            SoftLayer_Network_Storage.failbackFromReplicant, result)
         self.assert_called_with(
             'SoftLayer_Network_Storage',
             'failbackFromReplicant',
@@ -373,9 +377,9 @@ class BlockTests(testing.TestCase):
         mock.return_value = [{'id': 449494, 'name': 'dal09'}]
 
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
-        mock.return_value = [fixtures.SoftLayer_Product_Package.SAAS_PACKAGE]
+        mock.return_value = [SoftLayer_Product_Package.SAAS_PACKAGE]
 
-        mock_volume = copy.deepcopy(fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
+        mock_volume = copy.deepcopy(SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
         mock_volume['storageType']['keyName'] = 'PERFORMANCE_BLOCK_STORAGE'
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
@@ -389,7 +393,7 @@ class BlockTests(testing.TestCase):
             service_offering='storage_as_a_service'
         )
 
-        self.assertEqual(fixtures.SoftLayer_Product_Order.placeOrder, result)
+        self.assertEqual(SoftLayer_Product_Order.placeOrder, result)
 
         self.assert_called_with(
             'SoftLayer_Product_Order',
@@ -418,9 +422,9 @@ class BlockTests(testing.TestCase):
         mock.return_value = [{'id': 449494, 'name': 'dal09'}]
 
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
-        mock.return_value = [fixtures.SoftLayer_Product_Package.SAAS_PACKAGE]
+        mock.return_value = [SoftLayer_Product_Package.SAAS_PACKAGE]
 
-        mock_volume = fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME
+        mock_volume = SoftLayer_Network_Storage.STAAS_TEST_VOLUME
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
 
@@ -433,7 +437,7 @@ class BlockTests(testing.TestCase):
             service_offering='storage_as_a_service'
         )
 
-        self.assertEqual(fixtures.SoftLayer_Product_Order.placeOrder, result)
+        self.assertEqual(SoftLayer_Product_Order.placeOrder, result)
 
         self.assert_called_with(
             'SoftLayer_Product_Order',
@@ -463,7 +467,7 @@ class BlockTests(testing.TestCase):
             virtual_guest_ids=[200],
             ip_address_ids=[300])
 
-        self.assertEqual(fixtures.SoftLayer_Network_Storage.
+        self.assertEqual(SoftLayer_Network_Storage.
                          allowAccessFromHostList, result)
 
         self.assert_called_with(
@@ -478,7 +482,7 @@ class BlockTests(testing.TestCase):
             virtual_guest_ids=[200],
             ip_address_ids=[300])
 
-        self.assertEqual(fixtures.SoftLayer_Network_Storage.
+        self.assertEqual(SoftLayer_Network_Storage.
                          removeAccessFromHostList, result)
 
         self.assert_called_with(
@@ -491,7 +495,7 @@ class BlockTests(testing.TestCase):
             12345,
             subnet_ids=[12345678])
 
-        self.assertEqual(fixtures.SoftLayer_Network_Storage_Allowed_Host.
+        self.assertEqual(SoftLayer_Network_Storage_Allowed_Host.
                          assignSubnetsToAcl, result)
 
         self.assert_called_with(
@@ -504,7 +508,7 @@ class BlockTests(testing.TestCase):
             12345,
             subnet_ids=[12345678])
 
-        self.assertEqual(fixtures.SoftLayer_Network_Storage_Allowed_Host.
+        self.assertEqual(SoftLayer_Network_Storage_Allowed_Host.
                          removeSubnetsFromAcl, result)
 
         self.assert_called_with(
@@ -515,7 +519,7 @@ class BlockTests(testing.TestCase):
     def test_get_subnets_in_acl(self):
         result = self.block.get_subnets_in_acl(12345)
 
-        self.assertEqual(fixtures.SoftLayer_Network_Storage_Allowed_Host.
+        self.assertEqual(SoftLayer_Network_Storage_Allowed_Host.
                          getSubnetsInAcl, result)
 
         self.assert_called_with(
@@ -526,7 +530,7 @@ class BlockTests(testing.TestCase):
     def test_create_snapshot(self):
         result = self.block.create_snapshot(123, 'hello world')
 
-        self.assertEqual(fixtures.SoftLayer_Network_Storage.createSnapshot,
+        self.assertEqual(SoftLayer_Network_Storage.createSnapshot,
                          result)
 
         self.assert_called_with(
@@ -538,7 +542,7 @@ class BlockTests(testing.TestCase):
         result = self.block.restore_from_snapshot(12345678, 87654321)
 
         self.assertEqual(
-            fixtures.SoftLayer_Network_Storage.restoreFromSnapshot,
+            SoftLayer_Network_Storage.restoreFromSnapshot,
             result)
         self.assert_called_with(
             'SoftLayer_Network_Storage',
@@ -549,7 +553,7 @@ class BlockTests(testing.TestCase):
         result = self.block.enable_snapshots(12345678, 'WEEKLY', 10,
                                              47, 16, 'FRIDAY')
 
-        self.assertEqual(fixtures.SoftLayer_Network_Storage.enableSnapshots,
+        self.assertEqual(SoftLayer_Network_Storage.enableSnapshots,
                          result)
 
         self.assert_called_with(
@@ -560,7 +564,7 @@ class BlockTests(testing.TestCase):
     def test_disable_snapshots(self):
         result = self.block.disable_snapshots(12345678, 'HOURLY')
 
-        self.assertEqual(fixtures.SoftLayer_Network_Storage.disableSnapshots,
+        self.assertEqual(SoftLayer_Network_Storage.disableSnapshots,
                          result)
         self.assert_called_with(
             'SoftLayer_Network_Storage',
@@ -571,7 +575,7 @@ class BlockTests(testing.TestCase):
         result = self.block.list_volume_schedules(12345678)
 
         self.assertEqual(
-            fixtures.SoftLayer_Network_Storage.listVolumeSchedules,
+            SoftLayer_Network_Storage.listVolumeSchedules,
             result)
 
         expected_mask = 'schedules[type,properties[type]]'
@@ -585,16 +589,16 @@ class BlockTests(testing.TestCase):
 
     def test_order_block_snapshot_space_upgrade(self):
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
-        mock.return_value = [fixtures.SoftLayer_Product_Package.SAAS_PACKAGE]
+        mock.return_value = [SoftLayer_Product_Package.SAAS_PACKAGE]
 
-        mock_volume = copy.deepcopy(fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
+        mock_volume = copy.deepcopy(SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
         mock_volume['storageType']['keyName'] = 'ENDURANCE_BLOCK_STORAGE'
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
 
         result = self.block.order_snapshot_space(102, 20, None, True)
 
-        self.assertEqual(fixtures.SoftLayer_Product_Order.placeOrder, result)
+        self.assertEqual(SoftLayer_Product_Order.placeOrder, result)
 
         self.assert_called_with(
             'SoftLayer_Product_Order',
@@ -615,15 +619,15 @@ class BlockTests(testing.TestCase):
 
     def test_order_block_snapshot_space(self):
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
-        mock.return_value = [fixtures.SoftLayer_Product_Package.SAAS_PACKAGE]
+        mock.return_value = [SoftLayer_Product_Package.SAAS_PACKAGE]
 
-        mock_volume = fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME
+        mock_volume = SoftLayer_Network_Storage.STAAS_TEST_VOLUME
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
 
         result = self.block.order_snapshot_space(102, 10, None, False)
 
-        self.assertEqual(fixtures.SoftLayer_Product_Order.placeOrder, result)
+        self.assertEqual(SoftLayer_Product_Order.placeOrder, result)
 
         self.assert_called_with(
             'SoftLayer_Product_Order',
@@ -643,7 +647,10 @@ class BlockTests(testing.TestCase):
         )
 
     def test_order_block_replicant_os_type_not_found(self):
-        mock_volume = copy.deepcopy(fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
+        mock_package = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
+        mock_package.return_value = [SoftLayer_Product_Package.SAAS_PACKAGE]
+
+        mock_volume = copy.deepcopy(SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
         del mock_volume['osType']
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
@@ -665,9 +672,9 @@ class BlockTests(testing.TestCase):
         mock.return_value = [{'id': 449494, 'name': 'dal09'}]
 
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
-        mock.return_value = [fixtures.SoftLayer_Product_Package.SAAS_PACKAGE]
+        mock.return_value = [SoftLayer_Product_Package.SAAS_PACKAGE]
 
-        mock_volume = copy.deepcopy(fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
+        mock_volume = copy.deepcopy(SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
         mock_volume['storageType']['keyName'] = 'PERFORMANCE_BLOCK_STORAGE'
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
@@ -679,7 +686,7 @@ class BlockTests(testing.TestCase):
             os_type='XEN'
         )
 
-        self.assertEqual(fixtures.SoftLayer_Product_Order.placeOrder, result)
+        self.assertEqual(SoftLayer_Product_Order.placeOrder, result)
 
         self.assert_called_with(
             'SoftLayer_Product_Order',
@@ -712,15 +719,15 @@ class BlockTests(testing.TestCase):
         mock.return_value = [{'id': 449494, 'name': 'dal09'}]
 
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
-        mock.return_value = [fixtures.SoftLayer_Product_Package.SAAS_PACKAGE]
+        mock.return_value = [SoftLayer_Product_Package.SAAS_PACKAGE]
 
-        mock_volume = fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME
+        mock_volume = SoftLayer_Network_Storage.STAAS_TEST_VOLUME
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
 
         result = self.block.order_replicant_volume(102, 'WEEKLY', 'dal09')
 
-        self.assertEqual(fixtures.SoftLayer_Product_Order.placeOrder, result)
+        self.assertEqual(SoftLayer_Product_Order.placeOrder, result)
 
         self.assert_called_with(
             'SoftLayer_Product_Order',
@@ -749,9 +756,9 @@ class BlockTests(testing.TestCase):
 
     def test_order_block_duplicate_origin_os_type_not_found(self):
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
-        mock.return_value = [fixtures.SoftLayer_Product_Package.SAAS_PACKAGE]
+        mock.return_value = [SoftLayer_Product_Package.SAAS_PACKAGE]
 
-        mock_volume = copy.deepcopy(fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
+        mock_volume = copy.deepcopy(SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
         del mock_volume['osType']
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
@@ -767,9 +774,9 @@ class BlockTests(testing.TestCase):
 
     def test_order_block_duplicate_performance_no_duplicate_snapshot(self):
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
-        mock.return_value = [fixtures.SoftLayer_Product_Package.SAAS_PACKAGE]
+        mock.return_value = [SoftLayer_Product_Package.SAAS_PACKAGE]
 
-        mock_volume = copy.deepcopy(fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
+        mock_volume = copy.deepcopy(SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
         mock_volume['storageType']['keyName'] = 'PERFORMANCE_BLOCK_STORAGE'
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
@@ -778,7 +785,7 @@ class BlockTests(testing.TestCase):
             102,
             duplicate_snapshot_size=0)
 
-        self.assertEqual(fixtures.SoftLayer_Product_Order.placeOrder, result)
+        self.assertEqual(SoftLayer_Product_Order.placeOrder, result)
 
         self.assert_called_with(
             'SoftLayer_Product_Order',
@@ -804,9 +811,9 @@ class BlockTests(testing.TestCase):
 
     def test_order_block_duplicate_performance(self):
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
-        mock.return_value = [fixtures.SoftLayer_Product_Package.SAAS_PACKAGE]
+        mock.return_value = [SoftLayer_Product_Package.SAAS_PACKAGE]
 
-        mock_volume = copy.deepcopy(fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
+        mock_volume = copy.deepcopy(SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
         mock_volume['storageType']['keyName'] = 'PERFORMANCE_BLOCK_STORAGE'
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
@@ -820,7 +827,7 @@ class BlockTests(testing.TestCase):
             duplicate_snapshot_size=10
         )
 
-        self.assertEqual(fixtures.SoftLayer_Product_Order.placeOrder, result)
+        self.assertEqual(SoftLayer_Product_Order.placeOrder, result)
 
         self.assert_called_with(
             'SoftLayer_Product_Order',
@@ -848,9 +855,9 @@ class BlockTests(testing.TestCase):
 
     def test_order_block_duplicate_endurance_no_duplicate_snapshot(self):
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
-        mock.return_value = [fixtures.SoftLayer_Product_Package.SAAS_PACKAGE]
+        mock.return_value = [SoftLayer_Product_Package.SAAS_PACKAGE]
 
-        mock_volume = fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME
+        mock_volume = SoftLayer_Network_Storage.STAAS_TEST_VOLUME
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
 
@@ -858,7 +865,7 @@ class BlockTests(testing.TestCase):
             102,
             duplicate_snapshot_size=0)
 
-        self.assertEqual(fixtures.SoftLayer_Product_Order.placeOrder, result)
+        self.assertEqual(SoftLayer_Product_Order.placeOrder, result)
 
         self.assert_called_with(
             'SoftLayer_Product_Order',
@@ -883,9 +890,9 @@ class BlockTests(testing.TestCase):
 
     def test_order_block_duplicate_endurance(self):
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
-        mock.return_value = [fixtures.SoftLayer_Product_Package.SAAS_PACKAGE]
+        mock.return_value = [SoftLayer_Product_Package.SAAS_PACKAGE]
 
-        mock_volume = fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME
+        mock_volume = SoftLayer_Network_Storage.STAAS_TEST_VOLUME
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
 
@@ -898,7 +905,7 @@ class BlockTests(testing.TestCase):
             duplicate_snapshot_size=10
         )
 
-        self.assertEqual(fixtures.SoftLayer_Product_Order.placeOrder, result)
+        self.assertEqual(SoftLayer_Product_Order.placeOrder, result)
 
         self.assert_called_with(
             'SoftLayer_Product_Order',
@@ -925,16 +932,16 @@ class BlockTests(testing.TestCase):
 
     def test_order_block_modified_performance(self):
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
-        mock.return_value = [fixtures.SoftLayer_Product_Package.SAAS_PACKAGE]
+        mock.return_value = [SoftLayer_Product_Package.SAAS_PACKAGE]
 
-        mock_volume = copy.deepcopy(fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
+        mock_volume = copy.deepcopy(SoftLayer_Network_Storage.STAAS_TEST_VOLUME)
         mock_volume['storageType']['keyName'] = 'PERFORMANCE_BLOCK_STORAGE'
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
 
         result = self.block.order_modified_volume(102, new_size=1000, new_iops=2000, new_tier_level=None)
 
-        self.assertEqual(fixtures.SoftLayer_Product_Order.placeOrder, result)
+        self.assertEqual(SoftLayer_Product_Order.placeOrder, result)
         self.assert_called_with(
             'SoftLayer_Product_Order',
             'placeOrder',
@@ -948,15 +955,15 @@ class BlockTests(testing.TestCase):
 
     def test_order_block_modified_endurance(self):
         mock = self.set_mock('SoftLayer_Product_Package', 'getAllObjects')
-        mock.return_value = [fixtures.SoftLayer_Product_Package.SAAS_PACKAGE]
+        mock.return_value = [SoftLayer_Product_Package.SAAS_PACKAGE]
 
-        mock_volume = fixtures.SoftLayer_Network_Storage.STAAS_TEST_VOLUME
+        mock_volume = SoftLayer_Network_Storage.STAAS_TEST_VOLUME
         mock = self.set_mock('SoftLayer_Network_Storage', 'getObject')
         mock.return_value = mock_volume
 
         result = self.block.order_modified_volume(102, new_size=1000, new_iops=None, new_tier_level=4)
 
-        self.assertEqual(fixtures.SoftLayer_Product_Order.placeOrder, result)
+        self.assertEqual(SoftLayer_Product_Order.placeOrder, result)
         self.assert_called_with(
             'SoftLayer_Product_Order',
             'placeOrder',
@@ -976,4 +983,4 @@ class BlockTests(testing.TestCase):
 
     def test_list_block_volume_limit(self):
         result = self.block.list_block_volume_limit()
-        self.assertEqual(fixtures.SoftLayer_Network_Storage.getVolumeCountLimits, result)
+        self.assertEqual(SoftLayer_Network_Storage.getVolumeCountLimits, result)
